squashed backend migrations for polish
